+++ nifi_tools-processors/src/main/java/com/isgneuro/nifi/tools/bloom/BloomWithStopWatch.java
@@ -3,32 +3,23 @@ package com.isgneuro.nifi.tools.bloom;
 import org.apache.spark.util.sketch.BloomFilter;
 import org.apache.nifi.util.StopWatch;
 
+import java.util.HashSet;
 import java.util.Set;
 
 
 public class BloomWithStopWatch{
-    BloomFilter bloomFilter;
-    Set<String> tokens;
-    StopWatch stopWatch;
+    private BloomWithTokens bloomWithTokens;
+    private StopWatch stopWatch;
 
-    public BloomWithStopWatch(BloomFilter bloomFilter, StopWatch stopWatch) {
-        this.bloomFilter = bloomFilter;
+    public BloomWithStopWatch(BloomWithTokens bloomFilterWithTokens, StopWatch stopWatch) {
+        this.bloomWithTokens = bloomFilterWithTokens;
         this.stopWatch = stopWatch;
     }
-
-    public BloomFilter getBloomFilter() {
-        return bloomFilter;
+    public BloomWithTokens getBloomWithTokens() {
+        return bloomWithTokens;
     }
-
-    public void setBloomFilter(BloomFilter bloomFilter) {
-        this.bloomFilter = bloomFilter;
-    }
-
     public StopWatch getStopWatch() {
         return stopWatch;
     }
 
-    public void setStopWatch(StopWatch stopWatch) {
-        this.stopWatch = stopWatch;
-    }
 }